requirements.txt
Flask
rembg
Pillow
opencv-python-headless
numpy
